Remove dapi data from state, general clean up
